Tic Tac Toe/requirements.txt
pygame
psycopg2
python-dotenv
